Rename class to our project name
