+++ src/main/java/seedu/mtracker/MTracker.java
@@ -1,8 +1,8 @@
-package seedu.duke;
+package seedu.mtracker;
 
 import java.util.Scanner;
 
-public class Duke {
+public class MTracker {
     /**
      * Main entry-point for the java.duke.Duke application.
      */
