+++ src/test/java/seedu/mtracker/MTrackerTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.mtracker;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class MTrackerTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
